move updateOrAddCoin to types
