+++ precompiles/erc20/types.go
@@ -5,6 +5,7 @@ package erc20
 
 import (
 	"fmt"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -129,4 +130,22 @@ func ParseBalanceOfArgs(args []interface{}) (common.Address, error) {
 	}
 
 	return account, nil
+}
+
+// updateOrAddCoin replaces the coin of the given denomination in the coins slice or adds it if it
+// does not exist yet.
+//
+// CONTRACT: Requires the coins struct to contain at most one coin of the given
+// denom.
+func updateOrAddCoin(coins sdk.Coins, coin sdk.Coin) sdk.Coins {
+	for idx, c := range coins {
+		if c.Denom == coin.Denom {
+			coins[idx] = coin
+			return coins
+		}
+	}
+
+	// NOTE: if no coin with the correct denomination is in the coins slice, we
+	// add it here.
+	return coins.Add(coin)
 }
